requirements.txt
astroid==2.1.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
elasticsearch==6.3.1
greenlet==0.4.15
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.5
mccabe==0.6.1
motor==1.0
pycodestyle==2.4.0
pylint==2.2.2
python-dateutil==2.7.5
python-slugify==1.2.6
pytz==2018.7
requests==2.20.1
six==1.11.0
tornado==5.1.1
typed-ast==1.1.0
Unidecode==1.0.23
urllib3==1.24.1
wrapt==1.10.11
WTForms==2.2.1
wtforms-tornado==0.0.2
